requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
contourpy==1.1.1
cycler==0.12.1
diffusers==0.12.1
filelock==3.9.0
fonttools==4.50.0
fsspec==2024.3.1
huggingface-hub==0.21.4
idna==3.6
importlib_metadata==7.1.0
importlib_resources==6.3.2
install==1.3.5
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.7.5
mpmath==1.3.0
mypy-extensions==1.0.0
networkx==3.0
numpy==1.24.1
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==8.7.0.84
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-nccl-cu11==2.19.3
nvidia-nvtx-cu11==11.8.86
packaging==24.0
pillow==10.2.0
pkg_resources==0.0.0
pyparsing==3.1.2
pyrallis==0.3.1
python-dateutil==2.9.0.post0
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
safetensors==0.4.2
scikit-learn==1.3.2
scipy==1.10.1
six==1.16.0
sympy==1.12
threadpoolctl==3.4.0
tokenizers==0.15.2
torch==2.2.1+cu118
torchaudio==2.2.1+cu118
torchvision==0.17.1+cu118
tqdm==4.66.2
transformers==4.38.2
triton==2.2.0
typing-inspect==0.9.0
typing_extensions==4.8.0
urllib3==2.2.1
zipp==3.18.1
